Commented out sample-app volumes, updated blazegraph loading to so show incremental status, added stop option to keep dump files

+++ stop.sh
@@ -3,8 +3,9 @@
 
 
 usage () {
-  /bin/echo "Usage:  $0 [-c]"
-  /bin/echo "        -c      remove containers and downloaded dump files"
+  /bin/echo "Usage:  $0 [-c|w]"
+  /bin/echo "        -c      remove containers and downloaded dump files and mysql database"
+  /bin/echo "        -w      remove containers and mysql database, preserve downloaded dump files"
   exit 2
 }
 
@@ -13,32 +14,36 @@ clean_up () {
   echo "Removing docker containers"
   docker-compose down
 
-  if [ -f $data_dir/$mysql_file ] ; then
-    echo " "
-    echo "Removing mysql dump file"
-    rm "$data_dir/$mysql_file" 
-  fi
-  
-  # download may sometimes fail and create a directory
-  if [ -d $data_dir/$mysql_file ] ; then
-    echo " "
-    echo "Removing mysql dump file"
-    rmdir "$data_dir/$mysql_file" 
-  fi
+  # remove the dump files if -c option
+  if [ $cleanup -eq 1 ]; then
 
-  echo " "
-  for blazegraph_file in $blazegraph_models; do
-    if [ -f $data_dir/$blazegraph_file ] ; then
-      echo "Removing blazegraph import file $blazegraph_file"
-      rm "$data_dir/$blazegraph_file"
-    fi
-    # download may sometimes fail and create a directory
-    if [ -d $data_dir/$blazegraph_file ] ; then
+    if [ -f $data_dir/$mysql_file ] ; then
       echo " "
-      echo "Removing blazegraph import file $blazegraph_file"
-      rmdir "$data_dir/$blazegraph_file"
+      echo "Removing mysql dump file"
+      rm "$data_dir/$mysql_file" 
     fi
-  done
+    
+    # download may sometimes fail and create a directory
+    if [ -d $data_dir/$mysql_file ] ; then
+      echo " "
+      echo "Removing mysql dump file"
+      rmdir "$data_dir/$mysql_file" 
+    fi
+  
+    echo " "
+    for blazegraph_file in $blazegraph_models; do
+      if [ -f $data_dir/$blazegraph_file ] ; then
+        echo "Removing blazegraph import file $blazegraph_file"
+        rm "$data_dir/$blazegraph_file"
+      fi
+      # download may sometimes fail and create a directory
+      if [ -d $data_dir/$blazegraph_file ] ; then
+        echo " "
+        echo "Removing blazegraph import file $blazegraph_file"
+      rmdir "$data_dir/$blazegraph_file"
+      fi
+    done
+  fi
 
   if [ -f .env ] ; then
     echo " "
@@ -52,7 +57,7 @@ clean_up () {
       echo "Removing mysql database files"
       rm -r gridappsdmysql
     else
-      echo "Unable to remove gridappsdmysql, please run the following command."
+      echo "Error: unable to remove gridappsdmysql, please run the following command."
       echo "sudo rm -r gridappsdmysql"
     fi
   fi
@@ -64,10 +69,13 @@ data_dir="dumps"
 cleanup=0
 
 # parse options
-while getopts c option ; do
+while getopts cw option ; do
   case $option in
-    c) # Cleanup downloads and containers
+    c) # Cleanup downloads and containers and dump files
       cleanup=1
+      ;;
+    w) # Cleanup downloads and containers
+      cleanup=2
       ;;
     *) # Print Usage
       usage
@@ -80,7 +88,7 @@ echo " "
 echo "Shutting down the docker containers"
 docker-compose stop
 
-if [ $cleanup == 1 ]; then
+if [ $cleanup -gt 0 ]; then
   clean_up
 fi
